feat: add vercel serverless handler
