+++ api/[...route].ts
@@ -1,0 +1,8 @@
+import { createApp } from '../server/app';
+
+const appPromise = createApp();
+
+export default async function handler(req: any, res: any) {
+  const app = await appPromise;
+  return app(req, res);
+}
